chore: move ControlledJSONEditor
