+++ js/apps/system/_admin/aardvark/APP/react/src/components/jsonEditor/ControlledJSONEditor.tsx
